+++ CHANGELOG.md
@@ -39,6 +39,7 @@
 - Added `craft\services\Drafts::EVENT_AFTER_APPLY_DRAFT`.
 - Added `craft\services\Drafts::EVENT_BEFORE_APPLY_DRAFT`.
 - Added `craft\services\Drafts::publishDraft()`.
+- Added `craft\services\Globals::deleteSet()`.
 - Added `craft\services\Globals::reset()`.
 - Added `craft\test\ActiveFixture`.
 - Added `craft\test\DbFixtureTrait`.